// src/api/generated/models/examSummaryDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ExamSummaryDtoMaxLevel } from './examSummaryDtoMaxLevel';
import type { ExamSummaryDtoMinLevel } from './examSummaryDtoMinLevel';
import type { ExamSummaryDtoPurpose } from './examSummaryDtoPurpose';
import type { ExamSummaryDtoStatus } from './examSummaryDtoStatus';

export interface ExamSummaryDto {
  id?: string;
  code?: string;
  title?: string;
  shortDescription?: string;
  status?: ExamSummaryDtoStatus;
  purpose?: ExamSummaryDtoPurpose;
  minLevel?: ExamSummaryDtoMinLevel;
  maxLevel?: ExamSummaryDtoMaxLevel;
  versionNumber?: number;
  publishedAt?: string;
  updatedAt?: string;
}